Revert "Use tmux as TERM if TERM is screen but TMUX env is set"
Forcing TERM to tmux can cause problems, because often "tmux" terminfo is unavailable, which is why tmux defaults to "screen". The tmux FAQ explicitly mentions that tmux uses "screen" by default because it is widely available. This fixes the crash issues in gomuks, described in this bug: https://github.com/tulir/gomuks/issues/250

This reverts commit aef08b141333dfbaad8a7f0c187d7ab7d2b203c9.